# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FraudShield API - Financial Fraud Detection System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The channel the message came from
 */
export type AnalyzeRequestChannel =
  (typeof AnalyzeRequestChannel)[keyof typeof AnalyzeRequestChannel];

export const AnalyzeRequestChannel = {
  sms: "sms",
  email: "email",
  whatsapp: "whatsapp",
  other: "other",
} as const;

export interface AnalyzeRequest {
  /** The raw message text to analyze */
  message: string;
  /** The channel the message came from */
  channel: AnalyzeRequestChannel;
}

export type AnalysisResultClassification =
  (typeof AnalysisResultClassification)[keyof typeof AnalysisResultClassification];

export const AnalysisResultClassification = {
  safe: "safe",
  spam: "spam",
  fraud: "fraud",
} as const;

export type FraudIndicatorSeverity =
  (typeof FraudIndicatorSeverity)[keyof typeof FraudIndicatorSeverity];

export const FraudIndicatorSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface FraudIndicator {
  type: string;
  description: string;
  severity: FraudIndicatorSeverity;
}

export interface UrlCheck {
  url: string;
  isSuspicious: boolean;
  reason: string;
}

export interface AnalysisResult {
  id: number;
  message: string;
  channel: string;
  classification: AnalysisResultClassification;
  /** Confidence score 0-100 */
  confidence: number;
  /** Risk score 0-100 */
  riskScore: number;
  indicators: FraudIndicator[];
  detectedUrls: UrlCheck[];
  explanation: string;
  analyzedAt: string;
}

export interface AnalysisHistoryResponse {
  items: AnalysisResult[];
  total: number;
}

export interface DashboardStats {
  totalAnalyzed: number;
  fraudCount: number;
  spamCount: number;
  safeCount: number;
  fraudRate: number;
  urlsChecked: number;
  suspiciousUrls: number;
  avgRiskScore: number;
}

export interface RecentActivity {
  id: number;
  channel: string;
  classification: string;
  riskScore: number;
  messagePreview: string;
  analyzedAt: string;
}

export type ClassificationBreakdownByChannelItem = {
  channel: string;
  safe: number;
  spam: number;
  fraud: number;
};

export type ClassificationBreakdownByClassificationItem = {
  classification: string;
  count: number;
};

export type ClassificationBreakdownWeeklyTrendItem = {
  day: string;
  fraud: number;
  spam: number;
  safe: number;
};

export interface ClassificationBreakdown {
  byChannel: ClassificationBreakdownByChannelItem[];
  byClassification: ClassificationBreakdownByClassificationItem[];
  weeklyTrend: ClassificationBreakdownWeeklyTrendItem[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetAnalysisHistoryParams = {
  limit?: number;
  offset?: number;
};