# lib/api-zod/src/generated/types/analysisHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FraudShield API - Financial Fraud Detection System
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from "./analysisResult";

export interface AnalysisHistoryResponse {
  items: AnalysisResult[];
  total: number;
}