# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FraudShield API - Financial Fraud Detection System
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultClassification } from "./analysisResultClassification";
import type { FraudIndicator } from "./fraudIndicator";
import type { UrlCheck } from "./urlCheck";

export interface AnalysisResult {
  id: number;
  message: string;
  channel: string;
  classification: AnalysisResultClassification;
  /** Confidence score 0-100 */
  confidence: number;
  /** Risk score 0-100 */
  riskScore: number;
  indicators: FraudIndicator[];
  detectedUrls: UrlCheck[];
  explanation: string;
  analyzedAt: string;
}