# lib/api-zod/src/generated/types/analyzeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FraudShield API - Financial Fraud Detection System
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeRequestChannel } from "./analyzeRequestChannel";

export interface AnalyzeRequest {
  /** The raw message text to analyze */
  message: string;
  /** The channel the message came from */
  channel: AnalyzeRequestChannel;
}