# lib/api-zod/src/generated/types/classificationBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FraudShield API - Financial Fraud Detection System
 * OpenAPI spec version: 0.1.0
 */
import type { ClassificationBreakdownByChannelItem } from "./classificationBreakdownByChannelItem";
import type { ClassificationBreakdownByClassificationItem } from "./classificationBreakdownByClassificationItem";
import type { ClassificationBreakdownWeeklyTrendItem } from "./classificationBreakdownWeeklyTrendItem";

export interface ClassificationBreakdown {
  byChannel: ClassificationBreakdownByChannelItem[];
  byClassification: ClassificationBreakdownByClassificationItem[];
  weeklyTrend: ClassificationBreakdownWeeklyTrendItem[];
}