# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FraudShield API - Financial Fraud Detection System
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalAnalyzed: number;
  fraudCount: number;
  spamCount: number;
  safeCount: number;
  fraudRate: number;
  urlsChecked: number;
  suspiciousUrls: number;
  avgRiskScore: number;
}