# lib/api-zod/src/generated/types/recentActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FraudShield API - Financial Fraud Detection System
 * OpenAPI spec version: 0.1.0
 */

export interface RecentActivity {
  id: number;
  channel: string;
  classification: string;
  riskScore: number;
  messagePreview: string;
  analyzedAt: string;
}