# lib/api-zod/src/generated/types/urlCheck.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FraudShield API - Financial Fraud Detection System
 * OpenAPI spec version: 0.1.0
 */

export interface UrlCheck {
  url: string;
  isSuspicious: boolean;
  reason: string;
}